raspberry/easydispatch-collector/requirements.txt
PyYAML>=6.0
requests>=2.31.0
pyserial>=3.5

# Optional: OLED display support (for MMDVM status display)
# Install with: pip install luma.oled
# Requires: python3-dev, python3-pil, libfreetype6-dev, libjpeg-dev
luma.oled>=3.12.0
